fix: update imports for new ts-config
