+++ src/auth/controller/auth-controller.ts
@@ -1,6 +1,6 @@
 import {Request, Response, NextFunction} from 'express';
-import {AuthDAOController} from '../dao/auth-dao-controller.js';
-import {AuthorizedRequest} from '../../nverse/interface/authorized-request.js';
+import {AuthDAOController} from '../dao/auth-dao-controller';
+import {AuthorizedRequest} from '../../nverse/interface/authorized-request';
 
 export class AuthController {
