+++ src/auth/handler/auth-handler.ts
@@ -1,11 +1,11 @@
 import {Request, Response, NextFunction} from 'express';
-import {Tenant} from '../../tenant/interface/tenant.js';
-import {USER_ROLE} from '../../tenant/enum/user-role.js';
-import {UserRole} from '../../tenant/interface/user-role.js';
-import {LocalGatekeeper} from '../../nverse/local-gatekeeper.js';
-import {NewTenantValidator} from '../../nverse/validator/new-tenant-validator.js';
-import {TenantSanitizer} from '../../nverse/sanitizer/tenant-sanitizer.js';
-import {TenantDaoController} from '../../tenant/dao/tenant-dao-controller.js';
+import {Tenant} from '../../tenant/interface/tenant';
+import {USER_ROLE} from '../../tenant/enum/user-role';
+import {UserRole} from '../../tenant/interface/user-role';
+import {LocalGatekeeper} from '../../nverse/local-gatekeeper';
+import {NewTenantValidator} from '../../nverse/validator/new-tenant-validator';
+import {TenantSanitizer} from '../../nverse/sanitizer/tenant-sanitizer';
+import {TenantDaoController} from '../../tenant/dao/tenant-dao-controller';
 import {NVerseAuthorityResolver} from 'bmx-nverse-ts';
 import {getEntityHandler, unAuthGetEntityHandler, unAuthPostEntityHandler} from 'bmx-behemoth-ts';
 import {alfredLog} from 'bmx-alfred-ts';
