+++ src/constants.ts
@@ -1,4 +1,3 @@
 export class Constants {
-
 	public static readonly T: string = '';
 }
