+++ src/nverse/interface/authorized-request.ts
@@ -1,5 +1,5 @@
 import {Request} from 'express';
-import {Tenant} from '../../tenant/interface/tenant.js';
+import {Tenant} from '../../tenant/interface/tenant';
 
 export interface AuthorizedRequest extends Request {
 	tenant: Tenant;
