+++ src/nverse/sanitizer/tenant-sanitizer.ts
@@ -1,4 +1,4 @@
-import { Tenant } from "../../tenant/interface/tenant.js";
+import { Tenant } from '../../tenant/interface/tenant';
 import {NVerseSanitizer} from 'bmx-nverse-ts';
 
 export class TenantSanitizer extends NVerseSanitizer<Tenant, Tenant> {
